Update full_name.py
Improved on my structure for my comments based on recommendations of reviewer:

"Comments were next to the code, it is better to add it in a new line above the code you want to comment, keep a blank line above the comment also.

This improves the overall readability of your code and is essential for future maintenance and collaboration on projects. There are guidelines on how to structure your code and comments in Python known as PEP8 Guidelines. You can take a look at the official site for a more information on the PEP8 guidelines. https://peps.python.org/pep-0008/"