Update finance_calclators.py
Improved code based on reviewer reccomendations:

"While the effort to account for unusual inputs such as '10K' is appreciated. Please keep in mind that it is highly unusual for anyone to expect this kind of functionality so you are creating a program with more moving parts and that takes up more resources to cater to probably 1% of your potential users.
Your while loop would have been better implemented in making the calculator be able to perform more than one calculation per use case.

Please do not pad individual conditions in a set of conditional statements with blank lines. This doesn't help with readability and it takes up space

Please be mindful of using the continue keyword where it is redundant"